[mullet m2] Make expand children control visibility dynamic.
This CL is a part of an effort to better understand users
behavior with newly introduced field by field filling functionality.
It is not planned to be launched and will be removed after
the experimentation is done.

Under a feature flag parameter, this CL makes the expand child
suggestions control (currently always visible) visible if
there is any cell (of the corresponding suggestion) selected only.

https://screenshot.googleplex.com/9CLvCMV3AfMeQwE

Bug: 40274514
Change-Id: I35fcc544a62aa954c74ae82793f195e37ca1513d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5600301
Reviewed-by: Florian Leimgruber <fleimgruber@google.com>
Reviewed-by: Bruno Braga <brunobraga@google.com>
Commit-Queue: Dmitry Vykochko <vykochko@google.com>
Reviewed-by: Theo Cristea <theocristea@google.com>
Cr-Commit-Position: refs/heads/main@{#1311872}
NOKEYCHECK=True
GitOrigin-RevId: 22c82ac1803ffd60b69fc3d1a94c8b679486ffab