Optimize AssociateWindows on Android.
Placeholder tabs are significantly expensive to restore and try to
sync. We need a WebContents to do this correctly. If this is one time
expensive, that's okay. The problem is when a tab repeatedly fails to
restore, we end up paying this cost over and over.

This CL creates a default off feature, that when enabled, will try to
mitigate this inefficiency. The only known scenario where this is
really harming us is chrome internal urls, such as ntp or downloads.
Before we pay the price of creating an empty WebContents, we'll
deserialize the navigation entries, and verify that at least one of
the navigation entries seems to be syncable.

This initial check should be a subset of the true functionality that
decides if a tab is syncable. No change should visible externally,
the same tabs should be synced or resynced. But less wasted cycles.

It takes somewhere around 5-10ms to create an empty WebContents,
depending on the strength of the device. This was being done on start
up and every time a WebContents was modified/shown. There should be
significant performance improvements as a result when the feature is
enabled.

Bug: 347990994
Change-Id: I803a52819296e715f0e445f0ff085e50eee397d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5649069
Reviewed-by: Calder Kitagawa <ckitagawa@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Sky Malice <skym@chromium.org>
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1323054}
NOKEYCHECK=True
GitOrigin-RevId: 6ed9ac07624d05c3ed79dd4708b7fce3981d805e