Amend android-webview-code-coverage-native waterfall
Exclude tests not run by android-code-coverage-native.
This is to make it run only the WV-specific subset of the latter, as was the original intent; as well as an attempt to avoid linker "file too large" errors.

Change-Id: I2c62aa402b84986daf6417722bcbc822505d488a
Bug: None
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5683368
Auto-Submit: Mateusz Zbikowski <zbikowski@google.com>
Commit-Queue: Haiyang Pan <hypan@google.com>
Reviewed-by: Haiyang Pan <hypan@google.com>
Cr-Commit-Position: refs/heads/main@{#1325096}
NOKEYCHECK=True
GitOrigin-RevId: a186498da964491d36d7becb5d0b74d7982def3d