Add testing config for toolbar pinning and update Addresses action item
The addresses action item was causing test failures with the testing
config due to changes that had been made affecting the user decision
when the bubble was closed. This cl removes those changes, update the
action item to explicitly close the bubble, removes unneeded tests, and
adds to an existing test. This cl also updates the bubbles highlighted
view for all address bubbles so that the button anchors to and
highlights the correct button whether addresses is pinned or not.

Bug: 323962374, 352506087
Change-Id: Ieb98ce7a5475920998cff3764cf71f7d3edfe629
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5683527
Reviewed-by: David Pennington <dpenning@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Caroline Rising <corising@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1326125}
NOKEYCHECK=True
GitOrigin-RevId: 39af010cae6a0d3f26c59fe657b833aff8959ef3