[chromeperf] reset the shard size after the win-10-perf fleet is updated
This is a reland of https://chromium-review.googlesource.com/c/chromium/src/+/5714310

Reset the shard size of win-10-perf after fleet update.

Also relaxing the dimension value of OS, and removing the dimension of GPU. We have been relying on the synthetic product names and we are doing it on pinpoint.

Bug: b/344927054
Change-Id: Iae9fbb87e8ee8ea7111bcaa7962112e56c6131a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5714002
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>
Reviewed-by: John Chen <johnchen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1328984}
NOKEYCHECK=True
GitOrigin-RevId: db44fd3ccc0bd843f0fe5f52be151397e3f6d2f9