Clean up cast app menu experiment
This removes the experiment set up to evaluate cast usage in the app
menu. The decision is captured in the linked bug. When cast is available
the app menu becomes "Cast, save, and share" and cast it listed first in
the submenu.

Bug: 293176676
Change-Id: I4702494fac68ef3dbbd2bb86a6b1cdaa3d2643b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5717293
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: David Pennington <dpenning@chromium.org>
Commit-Queue: Caroline Rising <corising@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1330264}
NOKEYCHECK=True
GitOrigin-RevId: 7dcbac5813c3bec05031352daca25bf032c20951