Ablation Study: Introduce kBoostRenderProcessForLoadingPrioritizePrerendering feature param
When prerender is processed in the background, the process is
deprioritized. But the beacons are sent from the backgrounded page. This
CL prioritizes the renderer process that is used for prerendering. By
changing the priority, we can check how it affects the leakage.

Bug: 351953350
Change-Id: I2701ccb86aaca75f07a9900c7267018f948d8464
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5725436
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Commit-Queue: Minoru Chikamune <chikamune@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1330942}
NOKEYCHECK=True
GitOrigin-RevId: c535388d071f54759a0ccaa0674613673032a63d