Prompt API: fix built in AI blink features definition
In https://chromium-review.googlesource.com/c/chromium/src/+/5681534,
a new flag `kEnableBuiltInAIAPI` was added to control the built-in AI
registration logic like ProfiledKeyedService initialization and mojo
setup. The feature was declared as `implied_by` the prompt API, but
this implication is only effective for the blink feature checks.

In this CL, the base feature is marked as enabled by default, so the
chromium side will always have the necessary setup for built in AI API.

Bug: 354297775

Change-Id: I4fcbb1fedf8c6837392de7015cc61daa16f2be24
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5724922
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Commit-Queue: Mingyu Lei <leimy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1331492}
NOKEYCHECK=True
GitOrigin-RevId: acbb7f3623c5d7dfaee1d258632965c5a80857b1