[viz] Allow viz to ack upon activation when interactive
With this change, we introduce a mode in which we send acks to clients
when viz surfaces activate and we're "interactive" (that is, we're
within some number of frames of a frame where we had damage due to
scrolling or a touch interaction). We expect that this will cause some
extra / wasted work, but should help regularize the timing of the
client work since it effectively negates backpressure. That said, it
is still possible for timing to be irregular (i.e., it's not clear
how much this helps and this CL does not attempt to quantify this nor
check via integration test).

NOTE:
Earlier patch sets have the tests enabled by default to see which
tests would break and have attempted to address and parameterize
affected tests (see patchset 4).

credit: suggested by jonross@. Jon also noted the cause of the issue
with a previous attempt (crrev.com/c/5675451) -- that if a dependency
submitted early, it wouldn't early ack.

Change-Id: I1875780d946c2bf3304b6f43e428dd1271abeb7c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5683265
Commit-Queue: Ian Vollick <vollick@chromium.org>
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1333729}
NOKEYCHECK=True
GitOrigin-RevId: 2715769b7f7fe0f8a0f7663c11f2b48e43ab6e41