Run Crossbench on Android platform.
If the browser type starts with "android-" and an Android device is
attached, the requested browser APKs will be installed on the device and
the Crossbench test will be executed. The "--browser" value is replaced
with the "adb:<target browser's package name>".

Bug: 338630584
Change-Id: I3689d2c33c762e38e7289551bd2fd9e3b6a71707
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5599077
Reviewed-by: Wenbin Zhang <wenbinzhang@google.com>
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: Farid (Mojtaba) Faridzad <faridzad@google.com>
Cr-Commit-Position: refs/heads/main@{#1334461}
NOKEYCHECK=True
GitOrigin-RevId: 6edc75898618d6b9807a7d0a719079c3476cd91f