[infra] Add starlark declarations of the magic substitution args.
This adds placeholder structs that can be used wherever an argument is
expected. When generating .pyl files, the placeholder structs will be
replaced with the magic string value that generate_buildbot_json.py
recognizes to replace with the result of calling one of the functions
from //testing/buildbot/buildbot_magic_substituions.py.

The placeholders also hold a starlark function that will be used to
generate the resulting replacement arguments when a test is expanded in
starlark. Unfortunately, the logic must be duplicated between the two
locations because of location constraints imposed by angle and lucicfg.

Bug: 40258588
Change-Id: Ifb147caef4e5806246140840518840c2ff3b9205
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5750683
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Yuly Novikov <ynovikov@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Cr-Commit-Position: refs/heads/main@{#1335479}
NOKEYCHECK=True
GitOrigin-RevId: 29a454a742c77f2e1e4234033c7b95747a5e8810