Add parameter to delay the speculative RFH creation and filter out page restores
The CL adds a parameter to delay the speculative RFH creation so as not
to interfere with the loading.
The CL also adds a parameter to filter out page restores such as BFCache
restore and prerender activation. The same function
`GetFrameHostForNavigation` will be immediately called in
`OnResponseStarted`.

Bug: 332435024
Change-Id: Ic302ea2c3002f2e86f711fea2b0e65df1cf02f33
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5737130
Commit-Queue: Jiacheng Guo <gjc@google.com>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1335920}
NOKEYCHECK=True
GitOrigin-RevId: 782c9ac895e9afca95064f93d0964d7033a39181