Remove flag AndroidHatsRefactor
The flag is was only used for turning on start-up survey in tests to test the survey infra structure. This functionality does not have to be a flag, and is replaced with a static boolean in this CL. This also ensured the survey does not show up in irrelevant tests, since the start up survey is enabled in fieldtrial_testing_config for local builds.

Regarding test coverage, we have multiple experiments running surveys on Android at the time. Those features on pre-release channel also serves as a good means to ensure the code is being ran.

Fixed: 346588557
Change-Id: I84d455de53789b79ee593f2aea9209854ff7f1db
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5753067
Auto-Submit: Wenyu Fu <wenyufu@chromium.org>
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Commit-Queue: Theresa Sullivan <twellington@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1336597}
NOKEYCHECK=True
GitOrigin-RevId: 0637de6ef6f3b97af62838319a9df481bc53b934