TabDragging tests: Pass window hint to ui_controls more often
There were still a few calls to ui_controls left that didn't pass a
window hint but were used in test cases with multiple browser windows.
On Wayland without screen coordinates (i.e. Linux Wayland), this can
lead to the wrong window receiving the emulated mouse event requested
via ui_controls.

This CL adds DragInputToCenterNotifyWhenDone() as an alternative to
DragInputToNotifyWhenDone() that passes the correct window hint. Using
it also allows for a few small code cleanups.

We also need to rewrite the tests that rely on DragTabAndNotify() /
QuitDraggingObserver.Wait() returning immediately when attaching to a
new context, even when the tab dragging session continues running, as
that doesn't work with fallback tab dragging. Instead, we move the code
that ends the drag session into an extra step, and introduce
TabDragControllerInactiveWaiter to wait for the session to end so we can
verify our expectations for the final state.

With these changes, the DragToSeparateWindow,
DragToSeparateWindowDuringDragEnd, and DragAllToSeparateWindow test
cases now consistently pass on Wayland, regardless of whether fallback
tab dragging is enabled or not. (For fallback tab dragging we also need
https://crrev.com/c/5692091 for the tests to pass.)

Finally, with https://crrev.com/c/5538283 merged we can drop the check
whether a system-DnD-based tab dragging session is running in
ReleaseInputAfterWindowDetached() (which means all tests are now
agnostic of whether we use a system DnD session are not!).

Bug: 40238145
Change-Id: I23d40ae14cc314a5b1d2fa8204fa4ef4ae4b5bfb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5694567
Reviewed-by: Taylor Bergquist <tbergquist@chromium.org>
Commit-Queue: Max Ihlenfeldt <max@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1342747}
NOKEYCHECK=True
GitOrigin-RevId: fa49f3a9b2da2e925c8a7155a8ebf6233e816cbd