Add support for removing mixins when setting tests in starlark.
This change enables bundles to specify remove_mixins for a test to
prevent the named mixin from being able to be applied to the test by
that bundle or any bundle containing the bundle. It does so by using the
targets.per_test_modification struct, which allows for specifying mixins
or remove_mixins. The per_test_modifications continues to support using
a targets.remove object or specifying an individual mixin or list of
mixins. This support extends to legacy basic suites when they are used
as bundles.

This additionally implements the functionality of OS-specific args or
swarming details.

Bug: 40258588
Change-Id: I3e69a4835d8861851cde8d418f01e284fdb1b9fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5789645
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Cr-Commit-Position: refs/heads/main@{#1344237}
NOKEYCHECK=True
GitOrigin-RevId: ebd9995d468564a78d1454c2fb8bbe45351f1a9a