[infra] Provide gn -> ninja mapping for traffic_annotation_proto.
The linux-chromeos-annotator-rel builder specifies the script test
test_traffic_annotation_auditor, which returns the
traffic_annotation_proto target as a compile target. This works for for
the CI builder because the target simply gets added to the list of ninja
targets to build, but on the try builder, this fails during analyze
because dependency analysis happens in GN and gn_isolate_map.pyl doesn't
contain a mapping for traffic_annotation_proto. Without that, mb.py
analyze can't convert the ninja target to the label to pass to GN.

Bug: 40890977
Change-Id: I8999c076800fb01f54db46ed21325903af7ca683
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5805380
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1347425}
NOKEYCHECK=True
GitOrigin-RevId: b418e12b821733d663866afcbf1540b78a7e7b90