[performance_manager] Fix background title/favicon update scoring in BackgroundTabLoadingPolicy.
A restored WebContents does not have a URL (visible or committed)
when its PageNode is created [1]. As a result, the "visible URL"
passed to the constructor of PageNode is empty, no SiteDataReader
is associated with the PageNode until a navigation is actually
committed [2] and the "background title/favicon update" property
is not actually taken into account by BackgroundTabLoadingPolicy [3].

This CL fixes the issue by introducing
PageNodeImpl::SetMainFrameRestoredState() and calling it when
PageNodes are added BackgroundTabLoadingPolicy. At this point,
restored navigation entries have been added to the WebContents so the
URL is not empty (see CHECK in code). Setting the main frame URL
invokes SiteDataNodeData::OnMainFrameUrlChanged() and causes a
SiteDataReader to be associated with the PageNode.

Note: By default, Chrome uses
resource_coordinator::SessionRestorePolicy
instead of BackgroundTabLoadingPolicy to load background restored
tabs. This policy is affected by a similar issue: the main frame URL
has never changed so no SiteDataReader is associated with the PageNode
when TabDataAccess::SetUsedInBgFromSiteDataDB attempts to access it
(see proof: crrev.com/c/5813669).

[1] Restored WebContents and PageNode are created from
https://source.chromium.org/chromium/chromium/src/+/main:chrome/browser/ui/browser_tabrestore.cc;l=78-80;drc=82dff63dbf9db05e9274e11d9128af7b9f51ceaa
while restored navigation entries are added from
https://source.chromium.org/chromium/chromium/src/+/main:chrome/browser/ui/browser_tabrestore.cc;l=96-97;drc=82dff63dbf9db05e9274e11d9128af7b9f51ceaa
[2] https://source.chromium.org/chromium/chromium/src/+/main:components/performance_manager/decorators/site_data_node_data.cc;l=56;drc=82dff63dbf9db05e9274e11d9128af7b9f51ceaa
[3] https://source.chromium.org/chromium/chromium/src/+/main:chrome/browser/performance_manager/policies/background_tab_loading_policy.cc;l=408;drc=571289afa629784bb72b08312df7a2d547894827
[4] https://source.chromium.org/chromium/chromium/src/+/main:chrome/browser/resource_coordinator/session_restore_policy.cc;l=153-158;drc=82dff63dbf9db05e9274e11d9128af7b9f51ceaa

Bug: 40121561
Change-Id: I93863b7bd9e7cda572c13a0251af7a2bff1993dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5809762
Reviewed-by: Patrick Monette <pmonette@chromium.org>
Commit-Queue: Francois Pierre Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1347618}
NOKEYCHECK=True
GitOrigin-RevId: 48d1724faa9110a7134d4c0ab1977a9ae51471db