[PWA] Move link capturing flag to content/, rename to nav capturing
This CL moves the link capturing flag and related parameters to be
publicly accessible from the content/ layer. This allows the flag to
be used to control the behavior of the web contents implementation
introduced in crrev.com/c/5791853, and turned off if needed in case
some emergency.

Also renames the flag to be called PwaNavigationCapturing so that
it makes some sense when referred to from WebView, as done in
crrev.com/c/5809221. Renames mostly include the feature flag itself,
its usages as well as the strings to be shown on chrome://flags.

See context in https://chromium-review.googlesource.com/c/chromium/src/+/5791853/comment/825274d4_02b95051/

Bug: 362123239, 359599878, 351775835
Change-Id: I061f6cc6ab8998c87762c1cd6feab95767806ed3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5808589
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1348073}
NOKEYCHECK=True
GitOrigin-RevId: 9f0e9709a8d56053d381f44ed483d4e144469902