Update test_traffic_annotation_auditor to support not updating sheet.
The test_traffic_annotation_auditor test fails when running on a CrOS or
Windows try builder because it attempts to update the annotation sheet
which results in a PERMISSION_DENIED error. This is expected and
desirable since we wouldn't want a builder running against unreviewed
code to be able to make such changes.

The try build would succeed in this case because the builder will
attempt to run the test without the patch and encounter the same
failure, so it will exonerate the original failure. However, it takes a
non-negligible amount of time to do so. Providing a mechanism to skip
updating the sheet would avoid wasting this time.

This makes it so that a --no-update-sheet flag can be passed through the
script's --args flag. If it is present, then it will not attempt to
update the sheet.

Bug: 40890977
Change-Id: I4e1fda551f0d17d59165b5a2772e5acc7b36af6c
Cq-Include-Trybots: luci.chromium.try:linux-annotator-rel,linux-chromeos-annotator-rel,win-annotator-rel
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5819575
Reviewed-by: Nicolas Ouellet-Payeur <nicolaso@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Cr-Commit-Position: refs/heads/main@{#1348954}
NOKEYCHECK=True
GitOrigin-RevId: b7694dcf80c029418fc1631010c58397c93f0648