Revive WebGPUUseTintIR trial config for macOS
The _V2 version launched several months ago on ChromeOS, but we didn't
delete the config. We are now starting to Finch trial the same feature
on macOS, so I've reset the name and will add a new GCL config for
this trial.

Bug: 364675466
Change-Id: I6f1f5b712f979066836de2afde02f2bdbd4dfdce
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5847548
Commit-Queue: James Price <jrprice@google.com>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1353252}
NOKEYCHECK=True
GitOrigin-RevId: c067ebf851036a69a39003f478833fe0f9d0f227