Remove //testing/buildbot/tryserver.chromium.chromiumos.json
The only builder in this file has since been deleted, and the presubmit
checks don't appear to enforce the removal of a file once it's no longer
being generated.

So this removes the file and also the registration of the
lacros_all_tast_tests suite since nothing is running it any longer.

Bug: None
Change-Id: I7105868f7a4cc0142506fe706626a29fc5fb4959
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5852431
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Cr-Commit-Position: refs/heads/main@{#1353638}
NOKEYCHECK=True
GitOrigin-RevId: 7014d91342451d2470a909b7791382aa2d536601