[infra] Enable the use of variants in starlark.
This enables the use of variants by builders that set their tests in
starlark. This allows the use of matrix compound suites from starlark. A
bundle that uses variants can be declared by setting the variants
argument in the targets.bundle call, which will apply the variant to all
tests contained in the bundle.

Bug: 40258588
Change-Id: Ie3dc99ed9a06f32dd5b5d962196f059a6f417984
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5854503
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Cr-Commit-Position: refs/heads/main@{#1355365}
NOKEYCHECK=True
GitOrigin-RevId: b42c037fe488a9aafc21fa6f1b7b1f89e9c00db8