Add field trial testing config for SafeBrowsingSyncCheckerCheckAllowlist
It also captures a bug that the result is not delivered properly when
the URL is not on the allowlist or the blocklist. The callback should
be delivered via the complete check callback instead of the return
value. Added a unit test to capture this.

Bug: 347890373
Change-Id: Iae3c0f5bd60509614ef5d6ad73b1272792c67bf0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5857106
Commit-Queue: Xinghui Lu <xinghuilu@chromium.org>
Reviewed-by: Daniel Rubery <drubery@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1355414}
NOKEYCHECK=True
GitOrigin-RevId: 9353d6a38cae8893a4f10635e344d9c2c09beb8a