Reland "Add fieldtrial config for the Sensitive Content feature"
This is a reland of commit 33a5413a21e2d764dc379e45575cab67c9bdf269

The original CL causes test failures, because `mInternalsHolder` is null
inside `WebContentsImpl#getViewAndroidDelegate`.

Why `mInternalsHolder` is null is unclear, as the stack trace doesn't
provide much information about the original call site for the creation
of the tab and about why the tab's `WebContents` has a null
`mInternalsHolder`.
Stack trace: https://paste.googleplex.com/5449631275417600

This CL adds a null check to `mInternalsHolder`, with the promise to
immediately start investigating the root cause of the problem and fix it
if necessary. The null check is a temporary fix, so that we can start
experimenting with the feature in M130.

The failures caused by the original CL:
https://ci.chromium.org/ui/p/chromium/builders/ci/android-15-x64-rel/2581

Original change's description:
> Add fieldtrial config for the Sensitive Content feature
>
> Bug: 343119998
> Change-Id: I4c7fa55645bf034e10b504f802906581721a74b9
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5860616
> Reviewed-by: Jan Keitel <jkeitel@google.com>
> Commit-Queue: Theo Cristea <theocristea@google.com>
> Cr-Commit-Position: refs/heads/main@{#1355158}

Bug: 343119998
Change-Id: Idb3e026506a62183e8b4e5b759dd1bd1fb49081c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5865369
Reviewed-by: Jinsuk Kim <jinsukkim@chromium.org>
Commit-Queue: Theo Cristea <theocristea@google.com>
Cr-Commit-Position: refs/heads/main@{#1355854}
NOKEYCHECK=True
GitOrigin-RevId: 9c6be890deb031d9e36350d8c0eef847e1aa9342