Merge features with PerfCombined2024.
This will allow measuring the combined impact of the performance
optimizations.

Features being merged:
- LevelDBProtoAsyncWrite
- PumpFastToSleepAndroid
- ReduceCpuUtilization2

Note that ReduceCpuUtilization2 is not merged with the PerfCombined2024
on Webview because there was no independent field experiment for it on
Webview. We'll run that experiment first and then merge it with
PerfCombined2024.

Bug: 40287434
Change-Id: I8bbd5c15ccc77a7c2c60465a773061d1c0d91824
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5865614
Reviewed-by: Anthony Vallée-Dubois <anthonyvd@chromium.org>
Auto-Submit: Francois Pierre Doray <fdoray@chromium.org>
Commit-Queue: Anthony Vallée-Dubois <anthonyvd@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1355995}
NOKEYCHECK=True
GitOrigin-RevId: ea2be1041da57874b6081d0ac52cb6a116b4534d