domatolpm: introduce a webcodecs in_process fuzzer
This is a modified version of tiszka@ webcodecs domato grammar and template. This was only modified to fit better with how DomatoLPM works,
but the inner grammar remains unchanged.

Even though this fuzzer has already run for some time, the idea is both
to capitalize on the DomatoLPM implementation and use the existing
fuzzers in a coverage guided environment, in the hope that it could
improve generated testcases.

Change-Id: Ia3befcfb8bfa9cee9a312c9ed62a87643470dd6e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5841993
Reviewed-by: Adrian Taylor <adetaylor@chromium.org>
Commit-Queue: Paul Semel <paulsemel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1356407}
NOKEYCHECK=True
GitOrigin-RevId: db4823033eb81a0a11d3b12b7cc8cddc5cb71693