[iOS] Remove Finch flag for feed sync promo
The top-of-feed sync promo has been launched for a while and was
enabled by default in the client. This CL removes the Finch flag and
simplifies the feature code.

Converges SigninPromoViewStyleCompactHorizontal and SigninPromoViewStyleCompactVertical into a single SigninPromoViewStyleCompact, since we no longer need to toggle between
the compact layouts.

Change-Id: I9aa168af2a0ada1e09b8b26abdc91413e19bcf43
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5873526
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Commit-Queue: Adam Arcaro <adamta@google.com>
Cr-Commit-Position: refs/heads/main@{#1361876}
NOKEYCHECK=True
GitOrigin-RevId: ffd3657c807eba0c4b92a74c46a815cf4e6d4fc6