Update order variant details are applied to tests.
Details specified by variants and their mixins get applied before
anything else in generate_buildbot_json.py. This doesn't match the model
of starlark where bundles modify expansions produced by the bundles it
contains, so variants in a matrix compound suite would be applied after
details in a basic suite. This updates the application of variants so
that the variant details and its mixins are applied after applying the
mixins specified on a test in a basic suite.

Bug: 40258588
Change-Id: I9443032631f3611cb2c2d5f9986fc2b9af18829c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5900836
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Cr-Commit-Position: refs/heads/main@{#1362484}
NOKEYCHECK=True
GitOrigin-RevId: 65b7d36da7981185aa2429b52ad8d4c47b5c0c59