Handle test suite exceptions the same as mixins.
Currently generate_buildbot_json.py uses self.dictionary_merge to apply
changes in the modifications in test_suite_exceptions.pyl whereas mixins
use per-key logic for a handful of keys and then just uses the dict
update method to overwrite the entries for the remaining keys. The
dictionary_merge method recursively applies to dictionaries and when
merging two dicts, if the dict being merged in has None for a value,
that entry will be removed from the dict being merged into.

The different implementations result in swarming dimensions being
handled differently between a mixin and a test suite modification: a
dimension with a None value will be kept (written out as a json null) if
it was part of a mixin but the dimension will be removed if a test suite
exception sets a dimension to None. Since per test modifications in
starlark are handled via mixins, this creates an incompatibility for the
migration.

There are some instances of the null-value dimensions actually being
required since the recipe provides a default cpu dimension of x86-64 for
OSes besides CrOS and Android, so a null cpu dimension is required for
Fuchsia to remove the recipe-provided default. Additionally, the
resultdb field isn't merged despite it being a dict, so dictionary_merge
is ill-suited for handling the entirety of the mixins. To align the
behavior in both cases, the test suite exceptions will be applied as
though they were mixins, which aligns well with the starlark behavior.
This additionally removes the dictionary_merge method and replaces it
with a merg_swarming method to ensure that the other places that were
merging swarming dicts use the same behavior.

Bug: 40258588
Change-Id: I39b559c00dbdb2a09a6afa1aa5d6b35c6ae21890
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5902494
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Stephanie Kim <kimstephanie@google.com>
Cr-Commit-Position: refs/heads/main@{#1362708}
NOKEYCHECK=True
GitOrigin-RevId: 1b27162cad987260c004eb07ce19b9b052294ef7