Update fieldtrial config for CrOSLateBootSplitGenerations
The swappiness 60 group has bad performance and won't be launched, so
drop it. The 195 group seems the most promising, so move it to the first
spot so it gets tested. Also fix the names to match the Group.name
values, rather than the names of the Group variables.

Bug: None
Change-Id: I4b8e85a520e397cfa5b40b410d922e334440c4fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5882975
Auto-Submit: David Stevens <stevensd@chromium.org>
Reviewed-by: Charles Dick <cwd@google.com>
Commit-Queue: Charles Dick <cwd@google.com>
Cr-Commit-Position: refs/heads/main@{#1364059}
NOKEYCHECK=True
GitOrigin-RevId: 87997f6c43279acc1e7ab46e23e760747f80fb92