Decode JSON in Rust on main thread on Android
By default, DataDecoder defers JSON parsing to a thread pool, but for
the majority of cases the cost of scheduling the task far outweighs the
cost of parsing, so doing so in the current sequence makes sense. This
difference is particularly acute on Android platforms.

When using the C++ parser on Android, `DataDecoder::ParseJSON` sanitizes
and parses synchronously and invokes the callback in the same sequence.
This CL adds a feature parameter to use the Rust parser with the same
behavior, and a finch study will deploy this change on Android (enabling
the Rust parser) and on Desktop (which already uses the Rust parser, but
does so in a thread pool).

This is just a small step in a larger project. Subsequent changes will
remove JSON parsing from the data decoder entirely, replacing it in most
cases with a synchronous call to `base::JsonReader`.

Bug: 40276029
Change-Id: I5b53a8ba70855149e0708de304c90028f62b01d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5913185
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Dustin Mitchell <djmitche@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1366482}
NOKEYCHECK=True
GitOrigin-RevId: 01fd933ff0aacc46a5e2fa407b19ca602d34bc89